Move percentage width and height examples to the main width, height page.
Note when trying to use the CSS tesing component for the percentage
widths, the behaviour appears different than when using the same DOM and
CSS outside of CSS testing component. Specifically the elements that had
a percentage height of 100% with a parent who had a height of auto I
would have expected to change to a height of auto, but they ended up
stretching to the height of the iframe.

Used regular HTML/CSS for these examples instead for that reason.