Update inline and block page.
It was originally a page documenting a lot of the visual formatting model that related to inline and block. The visual formatting model page has since been a more thorough documenting of the whole visual formatting model. The inline and block page still had some useful live examples on it though. So removed what was duplicated from the visual formatting model page and left the live examples.

Also added some links from the visual formatting model page.

For some reason the normal flow example page anchors only seem to work when you're already on the page, but not when you go from another page.