Convert block % width and height examples to use CSS testing component.
Previously when using the component % height didn't work properly in the case of the containing block having a height of auto. You would expect the height of the element if a % to also be auto in this case. Instead it was taking the height of the nearest ancestor which had a fixed height. The fix for this was setting a doctype on the iframe document properly. Not prepending a doctype tag didn't seem to work, but writing the doctype tag as in this change did.