Viewport sizing - test fix for odd sizes on mobile.
Found that on mobile browsers on iOS 15.1 that different heights were
reported between these properties:
window.innerHeight,
document.documentElement.clientHeight

and these properties:
height: 100% (as child of height: 100% html and body tags)
height: 100vh

Removing border and padding on test elements to see if this fixes
things. box-sizing was set to border-box which in theory shouldn't have
this issue, but I am suspicious so testing. caniuse does say Safari 15.2
is where support is added to be fair...